start kmachine
