+++ game/kernel/jak1/kmachine.cpp
@@ -1,0 +1,120 @@
+#include "kmachine.h"
+
+#include <string>
+
+#include "game/kernel/common/fileio.h"
+#include "game/kernel/common/kboot.h"
+#include "game/kernel/common/kmachine.h"
+#include "game/kernel/common/kprint.h"
+
+namespace jak1 {
+
+/*!
+ * Initialize global variables based on command line parameters. Not called in retail versions,
+ * but it is present in the ELF.
+ * DONE
+ * Modified to use std::string, and removed call to fflush.
+ */
+void InitParms(int argc, const char* const* argv) {
+  // Modified default settings:
+  if (argc == 1) {
+    DiskBoot = 1;
+    isodrv = fakeiso;
+    modsrc = 0;
+    reboot = 0;
+    DebugSegment = 0;
+    MasterDebug = 0;
+  }
+
+  for (int i = 1; i < argc; i++) {
+    std::string arg = argv[i];
+    // DVD Settings
+    // ----------------------------
+
+    // the "cd" mode uses the DVD drive for everything. This is how the game runs in retail
+    if (arg == "-cd") {
+      Msg(6, "dkernel: cd mode\n");
+      isodrv = iso_cd;  // use the actual DVD drive for data files
+      modsrc = 1;       // use the DVD drive data for IOP modules
+      reboot = 1;       // Reboot the IOP (load new IOP runtime)
+    }
+
+    // the "cddata" uses the DVD drive for everything but IOP modules.
+    if (arg == "-cddata") {
+      Msg(6, "dkernel: cddata mode\n");
+      isodrv = iso_cd;  // tell IOP to use actual DVD drive for data files
+      modsrc = 0;       // don't use DVD drive for IOP modules
+      reboot = 0;       // no need to reboot the IOP
+    }
+
+    // the "deviso" mode is one of two modes for testing without the need for DVDs
+    if (arg == "-deviso") {
+      Msg(6, "dkernel: deviso mode\n");
+      isodrv = deviso;  // IOP deviso mode
+      modsrc = 0;       // no IOP module loading (there's no DVD to load from!)
+      reboot = 0;
+    }
+
+    // the "fakeiso" mode is the other of two modes for testing without the need for DVDs
+    if (arg == "-fakeiso") {
+      Msg(6, "dkernel: fakeiso mode\n");
+      isodrv = fakeiso;  // IOP fakeeiso mode
+      modsrc = 0;        // no IOP module loading (there's no DVD to load from!)
+      reboot = 0;
+    }
+
+    // an added mode to allow booting without a KERNEL.CGO for testing
+    if (arg == "-nokernel") {
+      Msg(6, "dkernel: no kernel mode\n");
+      MasterUseKernel = false;
+    }
+
+    // an added mode to allow booting without sound for testing
+    if (arg == "-nosound") {
+      Msg(6, "dkernel: no sound mode\n");
+      masterConfig.disable_sound = true;
+    }
+
+    // GOAL Settings
+    // ----------------------------
+
+    // the "demo" mode is used to pass the message "demo" to the gkernel in the DebugBootMessage
+    // (instead of play)
+    if (arg == "-demo") {
+      Msg(6, "dkernel: demo mode\n");
+      kstrcpy(DebugBootMessage, "demo");
+    }
+
+    // the "boot" mode is used to set GOAL up for running the game in retail mode
+    if (arg == "-boot") {
+      Msg(6, "dkernel: boot mode\n");
+      MasterDebug = 0;
+      DiskBoot = 1;
+      DebugSegment = 0;
+    }
+
+    // the "debug" mode is used to set GOAL up for debugging/development
+    if (arg == "-debug") {
+      Msg(6, "dkernel: debug mode\n");
+      MasterDebug = 1;
+      DebugSegment = 1;
+    }
+
+    // the "debug-mem" mode is used to set up GOAL in debug mode, but not to load debug-segments
+    if (arg == "-debug-mem") {
+      Msg(6, "dkernel: debug-mem mode\n");
+      MasterDebug = 1;
+      DebugSegment = 0;
+    }
+
+    // the "-level [level-name]" mode is used to inform the game to boot a specific level
+    // the default level is "#f".
+    if (arg == "-level") {
+      i++;
+      std::string levelName = argv[i];
+      Msg(6, "dkernel: level %s\n", levelName.c_str());
+      kstrcpy(DebugBootLevel, levelName.c_str());
+    }
+  }
+}
+}  // namespace jak1
