+++ game/kernel/jak1/kmachine.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace jak1 {
+/*!
+ * Initialize global variables based on command line parameters
+ */
+void InitParms(int argc, const char* const* argv);
+}  // namespace jak1
