+++ game/kernel/jak2/kboot.h
@@ -5,5 +5,8 @@
 #include "game/kernel/common/kboot.h"
 
 namespace jak2 {
+extern char DebugBootUser[64];
+extern char DebugBootArtGroup[64];
 s32 goal_main(int argc, const char* const* argv);
-}
+void kboot_init_globals();
+}  // namespace jak2
