+++ game/kernel/jak2/kmachine.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace jak2 {
+
+}
